[PIDM-1763] change amount type from Integer to Long (ecommerce-transactions-service) (#750)
* fix: change amount type from integer to long

* build: temporarily point commons dependency to feature branch via scm plugin

* chore: delete maven-scm-plugin

* chore: add maven-scm-plugin

* chore: update dep-sha

* chore: update dep-sha

* chore: update dep-sha

* chore: use 3.3.1-rc commons version

* chore: use 3.3.1-rc commons version

* chore: use 3.3.1-rc commons version

* chore: rollback 3.3.1 commons version

* chore: rollback 3.3.1 commons version

* chore: disable depcheck

* fix(PATCH auth-request): CHK-4809 handling patch auth request also in closure requested status (#749)

* handling patch auth request also in closure requested status

* fix tests

* fix tests

* remove ClosureRequestedProjectionHandler class, since it isn't anymore used

* remove transaction id from log text since it is present in MDC

* fix sonar issues

* Revert "fix sonar issues"

This reverts commit 1a84e77d7fc1d77a7e2fbb063de778716a70c4aa.

* Revert "remove ClosureRequestedProjectionHandler class, since it isn't anymore used"

This reverts commit e6a1530333d0edb71e7963f798b6cb5f41926e94.

* restore closureProjectionHandler and optimize closePayment pipeline

* Bump version [skip ci]

* chore: disable depcheck

* chore: disable depcheck

* chore: enable depcheck

* chore: disable depcheck

* chore: delete task in code-review pipeline

* test: use commons snapshot

* test: use commons snapshot 3.3.2-PIDM-1763-amount-int-to-long-SNAPSHOT-19278cb for test

* fix(tests): use Long literals for amount in TransactionRequestAuthorizationHandlerTest

* fix(tests): Fix Long handling in TransactionsUtilsTest assertions

* fix: set versione <pagopa-ecommerce-commons.version>3.5.0-PIDM-1763-amount-int-to-long-SNAPSHOT-917a202</pagopa-ecommerce-commons.version>

* fix: Fix commons version 3.5.0 and fix longValue in shouldGetTransactionTotalAmountV2 test

* fix: enable depcheck plugin

* fix: <pagopa-ecommerce-commons.version>3.6.0</pagopa-ecommerce-commons.version>

* fix: update dep-sha256.json

* test: test version commons 3.6.0-PIDM-1763-grandTotal-Long-SNAPSHOT-a7f6848

* test: test version commons 3.6.0-PIDM-1763-grandTotal-Long-SNAPSHOT-a7f6848

* fix: update commmons version 3.6.1

---------

+++ src/main/java/it/pagopa/transactions/client/PaymentGatewayClient.java
@@ -201,7 +201,7 @@ public class PaymentGatewayClient {
                                                             )
                                                     ) : null,
                                                     authorizationData.paymentNotices().stream()
-                                                            .mapToInt(
+                                                            .mapToLong(
                                                                     paymentNotice -> paymentNotice.transactionAmount()
                                                                             .value()
                                                             ).sum()
@@ -367,7 +367,7 @@ public class PaymentGatewayClient {
 
     private BigDecimal calculateGrandTotal(AuthorizationRequestData data) {
         long totalAmount = data.paymentNotices().stream()
-                .mapToInt(paymentNotice -> paymentNotice.transactionAmount().value())
+                .mapToLong(paymentNotice -> paymentNotice.transactionAmount().value())
                 .sum();
         return BigDecimal.valueOf(totalAmount + data.fee());
     }
@@ -497,7 +497,7 @@ public class PaymentGatewayClient {
                                             authorizationData
                                                     .paymentNotices()
                                                     .stream()
-                                                    .mapToInt(
+                                                    .mapToLong(
                                                             p -> p.transactionAmount().value()
                                                     )
                                                     .sum()
